feat: move screen to utils & rename home to dashboard in bottomNavItem
